requirements.txt
streamlit==1.29.0
pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
replicate==0.22.0

# Optional Azure dependencies (uncomment if using Azure storage)
# azure-storage-blob==12.19.0
# azure-identity==1.15.0 